requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==5.1
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.0
django-filter==24.3
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
et-xmlfile==1.1.0
idna==3.8
oauthlib==3.2.2
openpyxl==3.1.5
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
